refactor(adapters): reorganized module
moves to adapters module
