+++ src/allocation/adapters/repository.py
